+++ src/ship.rs
@@ -2,6 +2,8 @@ use bevy::prelude::*;
 use bevy_rapier2d::prelude::*;
 use leafwing_input_manager::prelude::*;
 
+use super::state::AppState;
+
 #[derive(Actionlike, PartialEq, Eq, Clone, Copy, Hash, Debug, Reflect)]
 pub enum ShipAction {
     Forward,
@@ -22,7 +24,10 @@ pub struct ShipPlugin;
 impl Plugin for ShipPlugin {
     fn build(&self, app: &mut App) {
         app.add_systems(Startup, setup);
-        app.add_systems(Update, ship_flight_system);
+        app.add_systems(
+            Update,
+            ship_flight_system.run_if(in_state(AppState::Active)),
+        );
     }
 }